Block 0: initial
This is a mostly finished block about a local environment that one would need to effectively follow the course.
It has a clean structure, and all necessary parts, so I sincerely hope it will undergo only cosmetic changes and additions in the future.
What I'm afraid the most now is the markup and links; I doubt I got it right at this moment. Markdown is still a skill to master to me.